canvas部分を別ファイルに分離
